requirements.txt
Django==2.1.5
djangorestframework==3.9.0
freezegun==0.3.11

whitenoise==3.3.1
gunicorn==19.7.1
dj-database-url==0.5.0
psycopg2
